+++ src/lib/st-label/st-label.component.spec.ts
@@ -9,11 +9,11 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Component, DebugElement } from '@angular/core';
+import { Component } from '@angular/core';
 
 // Component
 import { StLabelComponent } from './st-label.component';
-import { StTooltipModule } from '../st-tooltip/st-tooltip.module';
+import { StTooltipModule } from '../st-tooltip';
 
 @Component({
    template: `<label st-label></label>`
@@ -23,10 +23,8 @@ class TestStLabelComponent { }
 let component: TestStLabelComponent;
 let fixture: ComponentFixture<TestStLabelComponent>;
 let labelContent: string = 'label content';
-let labelId: string = 'label-id';
 let nativeElement: any;
 let template: string = '';
-let tooltipText: string = 'Label tooltip text';
 
 function createTestComponent(customTemplate?: string): Promise<ComponentFixture<TestStLabelComponent>> {
    if (customTemplate) {
@@ -42,8 +40,7 @@ function createTestComponent(customTemplate?: string): Promise<ComponentFixture<
 describe('StLabel', () => {
 
    beforeEach(async(() => {
-      TestBed
-      .configureTestingModule({
+      TestBed.configureTestingModule({
          declarations: [
             StLabelComponent,
             TestStLabelComponent
@@ -52,52 +49,24 @@ describe('StLabel', () => {
       });
    }));
 
-   it('Exist an element with class .content that contains label content', async(() => {
+   it('Exist an span element that contains label content', async(() => {
       template = '<label st-label>' + labelContent + '</label>';
       createTestComponent(template).then(() => {
          fixture = TestBed.createComponent(TestStLabelComponent);
          fixture.detectChanges();
          nativeElement = fixture.nativeElement;
-         expect(nativeElement.querySelector('.content').textContent).toContain(labelContent);
+         expect(nativeElement.querySelector('label').textContent).toContain(labelContent);
       });
    }));
 
-   it('Label content does not have an id if original tag does not', async(() => {
+   it('Label tag has st-label and st-tooltip classes', async(() => {
       template = '<label st-label>' + labelContent + '</label>';
       createTestComponent(template).then(() => {
          fixture = TestBed.createComponent(TestStLabelComponent);
          fixture.detectChanges();
          nativeElement = fixture.nativeElement;
-         expect(nativeElement.querySelector('#' + labelId + '-text')).toBeNull();
-      });
-   }));
-
-   it('Exist an element with same id as label plus sufix -text that contains label content', async(() => {
-      template = '<label st-label id="' + labelId + '">' + labelContent + '</label>';
-      createTestComponent(template).then(() => {
-         fixture = TestBed.createComponent(TestStLabelComponent);
-         fixture.detectChanges();
-         nativeElement = fixture.nativeElement;
-         expect(nativeElement.querySelector('#' + labelId + '-text').textContent).toContain(labelContent);
-      });
-   }));
-
-   it('Tooltip does not exists if property is not defined', async(() => {
-      createTestComponent().then(() => {
-         fixture = TestBed.createComponent(TestStLabelComponent);
-         fixture.detectChanges();
-         nativeElement = fixture.nativeElement;
-         expect(nativeElement.querySelector('.st-tooltip-content-text')).toBe(null);
-      });
-   }));
-
-   it('It has to display a text in its tooltip', async(() => {
-      template = '<label st-label id="' + labelId + '" tooltip="' + tooltipText + '">' + labelContent + '</label>';
-      createTestComponent(template).then(() => {
-         fixture = TestBed.createComponent(TestStLabelComponent);
-         fixture.detectChanges();
-         nativeElement = fixture.nativeElement;
-         expect(nativeElement.querySelector('.st-tooltip-content-text').textContent).toContain(tooltipText);
+         expect(nativeElement.querySelector('label').classList.contains('st-label')).toBeTruthy();
+         expect(nativeElement.querySelector('label').classList.contains('st-tooltip')).toBeTruthy();
       });
    }));
 });