+++ src/lib/st-switch/st-switch.module.ts
@@ -13,7 +13,6 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { StSwitchComponent } from './st-switch.component';
-import { StFormLabelModule } from '../utils/egeo-form/st-form-label/st-form-label.module';
 import { StLabelModule } from './../st-label/st-label.module';
 
 @NgModule({