+++ src/lib/st-tooltip/index.ts
@@ -8,15 +8,5 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-import { StFormLabelComponent } from './st-form-label.component';
-import { StTooltipModule } from '../../../st-tooltip/st-tooltip.module';
-
-@NgModule({
-   imports: [CommonModule, StTooltipModule],
-   declarations: [StFormLabelComponent],
-   exports: [StFormLabelComponent]
-})
-export class StFormLabelModule { }
+export * from './st-tooltip.component';
+export * from './st-tooltip.module';